[feature][api] 官方部署包安装 (#537)
1.新增 hdfs 和 yarn
2.安装hadoop包时自动创建hadoop软链

+++ datasophon-worker/src/main/java/com/datasophon/worker/strategy/resource/LinkStrategy.java
@@ -1,11 +1,14 @@
 package com.datasophon.worker.strategy.resource;
 
 
+import cn.hutool.core.io.FileUtil;
 import com.datasophon.common.Constants;
 import com.datasophon.common.utils.ShellUtils;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.extern.slf4j.Slf4j;
+
+import java.io.File;
 
 @Slf4j
 @EqualsAndHashCode(callSuper = true)
@@ -21,7 +24,13 @@ public class LinkStrategy extends ResourceStrategy {
     @Override
     public void exec() {
         String realTarget = basePath + Constants.SLASH + target;
-        ShellUtils.exceShell("ln -s " + source + " " + realTarget);
-        log.info("Create symbolic dir: {} to {}", source, realTarget);
+        File sourceFile = new File(source);
+        File targetFile = new File(realTarget);
+        if (!targetFile.exists() && sourceFile.exists()) {
+            // 先创建文件夹
+            FileUtil.mkdir(targetFile.getParent());
+            ShellUtils.exceShell("ln -s " + source + " " + realTarget);
+            log.info("Create symbolic dir: {} to {}", source, realTarget);
+        }
     }
 }
